remove reqs

+++ Makefile
@@ -76,13 +76,6 @@ shell: # Open a Django shell in the wagtail container
 bash: # Run bash in the wagtail container
 	$(wagtail) bash
 
-
-check-requirements: # Check whether requirements.txt needs re-generation based on pyproject.toml
-	$(wagtail-no-deps) poetry export --without-hashes | cmp -- requirements.txt -
-
-requirements: # Export the requirements to requirements.txt
-	$(wagtail-no-deps) poetry export --without-hashes --output ../requirements.txt
-
 clean: # Clean up python cache and webpack assets
 	npm run clean
 	find . -name '__pycache__' -exec rm -rf {} +
@@ -177,10 +170,11 @@ coverage: # Run tests with pytest and generate coverage report
 	$(testrunner) ../scripts/coverage.sh
 
 e2e-codegen: # Set up local environment and run Playwright's interactive test recorder
-	cp .env .env.orig
+	if [ ! -f .env.orig ]; then cp .env .env.orig; echo "backed up .env to .env.orig"; else echo "!! found existing .env.orig backup"; fi
 	cp .env.ci .env
 	docker compose stop wagtail
 	docker compose run --rm -d -p 8000:8000 --env DJANGO_SETTINGS_MODULE=config.settings.test --name wagtail-test-server wagtail
+	poetry run playwright install chromium
 	sleep 5
 	poetry run playwright codegen http://localhost:8000
 	mv .env.orig .env
